refactor employee management foundation

+++ src/main/java/com/practice/employee_ms/dto/CreateEmployeeRequest.java
@@ -7,6 +7,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -26,5 +28,5 @@ public class CreateEmployeeRequest {
     private String department;
 
     @Positive
-    private Integer salary;
+    private BigDecimal salary;
 }
